Allow `.default` to participate in common type determination
It makes the most sense to treat `.default` just like any other element in `values`. This means it must be size 1 or the same size as the logical conditions, and it means it should get a say in the common type determination. This change is motivated by the fact that we should be able to set `.default = missing` in `if_else()`, and `missing` should be part of the common type computation.